fix(conf_loader) do not apply additional processing when loading from `.kong_env` file
Doing so could cause failures such as ones described in #5761, where a
value of `"!abCDefGHijKL4\#1MN2OP3"` was first transformed to
`"!abCDefGHijKL4#1MN2OP3"` and written to `.kong_env`. If we apply
transformation to `"!abCDefGHijKL4#1MN2OP3"` again when loading from
`.kong_env`, then it becomes `"!abCDefGHijKL4` which results in
different value used by the CLI and worker.

fixes #5761