chore(makefile) make grpcurl part of dependencies
Our `make dev` "uninstall and then reinstall" phase makes it a bit
unfriendly to certain build systems (like pongo) which prefer to use
`make dependencies` instead.

One difference between `make dev` and `make dependencies` is
that the former installs grpcurl, but not the later.

This change moves the installation of grpcurl to `make dependencies`,
so it becomes "like `make dev` but without uninstalling and reinstalling".