fix(plugins) config.anonymous = "" in auth plugins
Fix https://github.com/Kong/kong/issues/5653

We could not configure an auth plugin with config.anonymous="" in CE 2.0
or EE 1.5.

It also broke users that had config.anonymous="" configured. To
reproduce, create an auth plugin with config.anonymous="" (empty string)
in 0.13, then upgrade all the way to 2.0.

This is the breaking change: https://github.com/Kong/kong/pull/5552

---

We have three options to fix it:

(a) a migration;
(b) start special-casing "" to mean null for this field in the code of
these plugins;
(c) add legacy back to those anonymous fields of string type fields,
meaning that in legacy fields "" auto-translates to null

This is the execution of (c).