fix(globalpatches) increase the limit when encoding sparse array in Kong
When Kong parses a declarative config, and when the declarative config
has a lot of entities and one of the high index entity has validation
errors, Kong may generate a JSON response message like this:

```
plugins = {
    [12] = {
        <schema errors>
    }
}
```

Unfortunately this causes the lua-cjson library to throw the
"Cannot serialise table: excessively sparse array" error. This PR
increases the limit of cjson so that these error won't happen with small
number of configurations.

fixes #5768