wip(autodoc) start adding missing info on admin-api (UNFINISHED)
This commit shows three different changes needed in the admin-api.lua file:
* Undocumented fields like `consumer.type`
* Fields we might want to skip docs about, like `plugins.run_on`
* Undocumented files like `kong/routes/admins.lua` (partially documented here)

On that last note: It seems that the file is performing "usual CRUD operations" with the Admins. If that's the case, I recommend considering removing `kong/routes/admins.lua` completely and leaving it automated. That way the only needed thing would be docummenting Admin as an Entity (like Services or Routes). The CRUD endpoints (with examples) would be generated automatically from it.