feat(balancer): expose balancer_health even when healthchecks are off
The Kong load balancer has concepts of upstream health, which is managed by
the healthchecker, and balancer health, which is affected by the healthchecker
but also affected by the balancer "health threshold" setting. This means a
_balancer_ can be unhealthy (and return HTTP 503) even when healthchecks for
the upstream are disabled.

The endpoint `/upstreams/<upstream>/health?balancer_health=1` is an "advanced
mode" for viewing health, which shows the _balancer health_ as opposed to
upstream health. (It is hidden behind a query argument to avoid confusion, as
for most cases the upstream health is sufficient and the "balancer"
terminology referring to a mostly internal Kong object may also be unclear to
end-users.) This endpoint, however, only returns the balancer health, when
healthchecks for upstream health are enabled.

This commit changes this behavior so that using `?balancer_health=1` we can always
see the balancer health, through a new attribute `balancer_health`, which
always returns `HEALTHY` or `UNHEALTHY` (reporting the true state of the
balancer), even if the overall upstream health status is `HEALTHCHECKS_OFF`.
This is useful for debugging.

The original attribute `health` is preserved with the existing semantics for
backward compatibility, returning `HEALTHY`, `UNHEALTHY` or
`HEALTHCHECKS_OFF`.

Signed-off-by: Aapo Talvensaari <aapo.talvensaari@gmail.com>