fix(schema) transformation validation is skipped when it should not
### Summary

This fixes issue when trying to merge master to next. The problem is
because `next` now has read-before-write on updates, but it also added
oauth2 credentials transformation and related tests. One of those tests
starts to fail if trying to merge master to next. The test itself is
fine.

The problematic test is this:
https://github.com/Kong/kong/blob/next/spec/03-plugins/25-oauth2/02-api_spec.lua#L522-L540

```lua
it("updating client_secret requires hash_secret", function()
  local res = assert(admin_client:send {
    method  = "PATCH",
    path    = "/consumers/bob/oauth2/" .. credential.id,
    body    = {
      client_secret = "test",
    },
    headers = {
      ["Content-Type"] = "application/json"
    }
  })
  local body = assert.res_status(400, res)
  local json = cjson.decode(body)
  assert.same({
    ["@entity"] = {
      "all or none of these fields must be set: 'hash_secret', 'client_secret'"
    }
  }, json.fields)
end)
```

The `status` code there is actually `200`, so the implicit
transformations validation is skipped in that case.

This commit fixes that issue and allows us then to merge master
to next without this test failing.