fix(schema) make process auto fields to remove empty strings on select
### Summary

On PR #5879 @murillopaula proposed a fix to a common but arbitrary plugin
field called `config.anonymous`. The PR tried to tackle two things at the
same time. It tried to infer `JSON` input (which we don't do), and it tried
to fix issue where upgraded databases still contained empty strings `""`
stored to `config.anonymous` (instead of `null`). This only fixes the issue
for this field. It also added `legacy` flag to that field, which in my
opinion the field is not, or that it is not any special to any other field
that could contain empty string `""`. So I didn't think it was a good idea
to start flagging fields to `legacy` one by one.

Thus I discussed this with @murillopaula and out of those discussions came
one winner that we both agreed upon. The solution is to run automatic
transformation on `select` to database fields that are not legacy, and
convert the value to `nil` (remove field) or `null` (when nulls were
requested). This is implemented directly in `process_auto_fields` which
we call automatically on daos. The solution looks like this:

```lua
if context == "select"
   and not field.legacy
   and field.type == "string"
   and (field.len_min or 1) > 0
   and data[key] == ""
then
  data[key] = nulls and null or nil
end
```

This fixes the issue for all the fields that are updated to new schema which
if not `len_min=0` will not accept `""` anymore as input, and thus are auto-
translated to `nil`/`null`.

### Issues resolved

Fix #5653 (the database `""`)
Close #5879 (previous attempt)