perf(basic-auth-v2) introduces basic-auth-v2 plugin with better performance
The original basic-auth plugin has id as the primary key, with a
secondary index on username. While this works well with postgres,
there are performance issues with cassandra as lookups always happen
on username and secondary indexes on cassandra are inefficient.

A new plugin is introduced as changing the primary key in cassandra
is non-trivial and hence updating the existing plugin is not possible.

This plugin makes username as the partition key and id as the
clustering in cassandra. In case of postgres, both form composite
primary key.

Signed-off-by: Abhishek Varshney <abhishek.varshney@razorpay.com>