chore(declarative) bump decl spec verno to 1.2
Releasing a new version of the 1.2 of the declarative spec file.

Changelog:

* New `_transform` field.

1.2 parser is backwards compatible in the sense that the new parser
understands and processes files with 1.1 syntax correctly (in that
case it will use the default, set to `true`, which is what was done
in 1.1). This is the reason why both options are allowed in the schema.

However, the 1.1 parser will error if it encounters the new
`_transform` option, hence the need of releasing a new
version.