refactor(declarative) change i/o params of dc
The main change this does is:

* The `flatten` method of the declarative config
  schema used to only return the entities, but not the config variables
  (like _format_version). Now it contains both (entities and config
  vars)

As a result, all the `declarative_config:parse*` methods return
a table with entities and config variables inside. If you want to
iterate over the entities only, you have to be careful to avoid the
conf vars. On the other hand, if you need both the entities and the
vars, they are all together in one place.

I also simplified the handling of the `err` and `err_t` variables
while I was there.

This needed in order to more easily add a new parameter to
declarative config on the next commit.