refactor(db-less) change the `flatten` return types
This PR modifies some of the changes introduced by #5835.

In particular, the `flatten` function now returns two separate
tables, `entities` and `meta`, instead of a single one.

This is done in order to is simplify a future merge with Kong EE.

The `_transform` field is also removed from the reports as it isn't
needed.