fix(dbless) actually release memory when purging the cache
This change makes it actually all items from the cache (not only
mark them as "expired"), so that `safe_set` used in a subsequent
declarative config load can work if the previous one failed with
"memory full" (due to a too-large configuration that returned
HTTP 413).

This fixes some flakiness previously observed in tests, where
tests would pass when running individually but fail when running
in sequence with a smaller `mem_cache_size` configuration.