feat(go) fake response phase
Adds a new 'response' phase, available for either Lua or Go plugins.
During this phase, the complete headers and body from the service's
response are availabe, and can be modified before sending back to the
client.

Handlers for this new phase are called at the end of OpenResty's
'access' phase, after the service request is performed via
`ngx.location.capture()`, and only if the buffered_proxy feature is
active.

Lua plugins that want to use this phase have to call
`kong.service.request.enable_buffering()` during the `access` phase to
enable this.  For Go plugins, the presence of a `Response` handler
automatically enables buffering.