fix(declarative) fix missing nulls on read
### Summary

Both Postgres and Cassandra strategies return null on fields that have
null value stored in database. The DBless stores entities directly to
cache where we don't keep nulls. But it has a problem that this PR
tries to fix. This PR changes off strategy to fill nulls when the
entities are accessed via DAO, the nulls are then later removed
(again) by DAO when it calls process_auto_fields. This extra fill
nulls is added to avoid the DAO process_auto_fields to fill explicit
null values with possible default values.

### Issues Resolved

Fix #5834