fix(schema) validate_primary_key to check for invalid input
### Summary

It was reported on #5974 that our code crashes with:

```
kong/db/schema/init.lua:1371: attempt to index local 'pk' (a number value)
```

On invalid input, in this case it was a declarative config with `number`
as pk instead of string.

This commit fixes the `schema:validate_primary_key` to check the input value
of `pk`.

### Issues Resolved

Fix #5974