feat(pdk) log serialization function `kong.log.serialize`
This PR makes the basic log serializer a proper PDK function `kong.log.serialize`. Calling the old module will generate a one time warning in the `error.log` file to remind developers to switch.

Background: the basic log serializer currently resides under the `plugin` directory, but it is not actually a proper plugin and logging plugins `require` the Lua file directly to use it.

This has created headache for plugin development, as the interface of the log serializer was never formalized and a lot of plugin user don't know how to use it. By making it a proper PDK function, it becomes officially part of the core interface, which makes it more suitable for use during plugin developments.

This PR also adds a new PDK function `kong.client.tls.set_client_verify` which allows overriding of the `request.tls.client_verify` value inside the log serializer output.

Supersedes #5995