fix(runloop) execute plugins' init_worker handler independently of plugin iterator build
On a multi-kong cluster, the plugins iterator can fail to build (if a worker changes it while another one is initializing).

When this happens on the init_worker phase, this prevents plugins from executing the `init_worker` callback on the plugins, leaving the worker on a bad state.

This change still tries to build the plugins iterator on the init_worker phase (so it is cached for future calls), but executes the init_worker callback of the plugins even if it fails.