fix(plugins) fix allow/deny nomenclature in access restriction plugins
Standardize on `allow` and `deny` as terms for access control in Kong plugins.

Previous nomenclature is deprecated and support will be removed in Kong 3.0.

Tests were adjusted for the new nomenclature; this PR was also tested with
the previous versions of the tests to ensure the deprecation system works.
All passed except for minor changes in error messages, which are not considered
a breaking change.