perf(balancer) use leader/followers workers on eventual
Execute costly code in only one worker (called the "leader"), the one
with PID 0). This way, only the leader will fetch the list of upstreams
from the database. Since we're leveraging mlcache, once the leader
performs the db query and fetches the list of upstreams from the L3
cache, the other workers (called the "followers") will retrieve this
list from the L2 cache. This model should be eventually consistent.