tests(migrations) tests for core/009
This commit aglutinates tests for both postgres and cassandra
strategies.

A couple problems were found and fixed with the tests:

In postgres, when the migration had already done the "up" action
but had yet to do the "teardown", an old node could create entities
with NULL in the ws_ids (this should not be allowed, they should have the
default ws_id instead). The fix involved moving the reset of the
default for that column to the teardown part of the migration.
(At which point, old nodes should already be shut down)

In cassandra, plugins cache-key was being applied to the wrong plugins
(instead of fixing the cache-keys which ended in ':', the migration
would fix the cache-keys which did *not* end in ':').

Both of these problems were fixed and extracted as hotfixes to our
2.1.0 branch.