fix(migrations) fix composite cache keys in Cassandra
There's a mismatch between the way PostgreSQL migrates
composite cache keys in Postgresql which end with ':':

```
UPDATE "$(TABLE)"
SET cache_key = CONCAT(cache_key, ':',
                       (SELECT id FROM workspaces WHERE name = 'default'))
WHERE cache_key LIKE '%:';
```

Cassandra did the opposite, updating the cache keys which did *not*
end with ':' :
```
for _, row in ipairs(rows) do
  if not row.cache_key:match(":$")
    local cql = ...
```

This change was masked because of the plugins cache prewarm, which
ignores the cache_key field. But it was visible if Kong was started
with cache prewarming deactivated for plugins, after migrating
plugins from an older version of kong in Cassandra.