fix(hybrid/vitals) add the missing cluster_telemetry_server_name and move cluster_telemetry_* to ee/conf_loader
`cluster_telemetry_server_name` will be used by the data plane to
specify SNI when connecting to control plane. It will be similar to
`cluster_server_name` for config sync and will fallback to
`cluster_server_name` if it's unset. The SNI will only be used when
`cluster_mtls` is set to `pki`.