fix(migrations) make cassandra ws_update_keys re-entrant
### Summary

When running migrations with `force`, e.g.:

```
kong migrations up -f
```

OR

```
kong migrations finish -f
```

Cassandra could ran `ws_update_keys` multiple times, even on a
row that was already updated. This commit makes it not to do that
for already updated rows.

It was noted that when migrations fail without `force` and they
are ran again without `force` there is re-entrancy problem too,
that this fixes.

This was added as a quick fix for upcoming 2.1.2.