fix(migrations) split cert_digest update to two loops
### Summary

In general, it is not good to modify a table/data that we are iterating
at the same time. This commit changes it so that first iterator collects
the statements, and second iteration then runs the updates.

This commit also makes it skip unneccessary updates, making it better on
re-entrancy scenarios.