remove outer worker lock
Remove the worker lock on calls to `declarative.load_into_cache_with_events()`.

There already was a "busy" early return, by checking the existence of the shared counter, but there was a significant gap between the check and the creation of the counter.  Replaced by a `DICT:add()`, which can (atomically) tell if the key was already there.

Add a `DICT:delete()` at any other error return that can happen before propagating the config event to other workers to avoid a spurious counter that would force false "busy" messages until expired.

Return an extra `ttl` value to send a slightly better retry delay to the admin client.