feat(openid-connect) id token is now preserved if a new one is not given when refreshing
### Summary

Originally we kept `id token` in session when refreshing the tokens even if the
refreshing didn't return id token (which they normally don't). This caused an
issue where the verification could fail if id token has expired. Thus on `0.2.8`
we removed id token. The previous commit in this PR changes it so that decoding
doesn't fail anymore with expired tokens.