chore(ci) use a stage for case
seems we are hitting concurrency issues when running
all the tests in parallel. Tried to setup Jenkins plugin
for concurrency but did not manage to make it work. This is
a temporary solution until we figure out a better one. The
bummer on this one is that once a step fails we cannot know
if the next ones work.