feat(hybrid) do not queue updates on dp nodes
### Summary

Currently when CP sends new config to DP the DP will apply that config
even in case it is already applying a config, and even in case when CP
sends a new config (which means 2 configs queued). Only the last config
is needed. This PR changes it so that DP drops the unneccessary configs
that it knows are going to be replaced right after.

This commit is continuation of the work that started with #6293 that
made the CP skip the configs that it knows are already going to be
replaced. This just makes the same thing on DP side.

You might wonder why is that? With huge configs, it can take a lot of
time for DP to apply the configuration, @cwurm told us delays like 77
seconds in this message: https://github.com/Kong/kong/pull/6280#issuecomment-684078073

That's why the CP can send multiple new configs during the time DP is
applying the previous ones. This commit is to fix that.