chore(runloop) just use one context value for buffered proxying
### Summary

Originally we implemented buffered proxying for `2.0` that had PDK function
`kong.service.request.enable_buffering()`. This enabled Lua plugins to read
the response body on `header_filter` phase but did not allow Go plugins to
do so as we cannot yield on `header_filter` and thus that phase is not available
for Go plugins.

We then decided to implement a fake `response` phase. This was implemented but
it added `ngx.ctx.has_response` context value that is not used anywhere else
than detecting whether to do buffered proxying or not. The mentioned `enable_buffering`
function already set `kong.ctx.core.buffered_proxying`. This commit changes so that
we have only one context value to detect that, so both of the above were removed and
now we detect buffered proxying by `ngx.ctx.buffered_proxying`. The reason why not
using `kong.ctx.core` for this is that in hot paths we already have `ngx.ctx` available
and thus it makes sense to use it.

This commit also clear the `ngx.ctx.buffered_proxying` in case it didn't really happen.