feat(core) a new response phase for buffered proxying (#5991)
* feat(go) fake response phase

Adds a new 'response' phase, available for either Lua or Go plugins.
During this phase, the complete headers and body from the service's
response are availabe, and can be modified before sending back to the
client.

Handlers for this new phase are called at the end of OpenResty's
'access' phase, after the service request is performed via
`ngx.location.capture()`, and only if the buffered_proxy feature is
active.

Lua plugins that want to use this phase have to call
`kong.service.request.enable_buffering()` during the `access` phase to
enable this.  For Go plugins, the presence of a `Response` handler
automatically enables buffering.

* any `response` handler sets buffering automatically

Check within existing plugin/phase loop in the plugin_iterator
constructor, avoid new check loop in init.response

(disable checks that contratict this. new tests pending)

* Reject plugins with both `response` and `_filter` methods

* tests: split test plugins with response and _filter

remove all tests that assume a single plugin can have both.

* check against legacy plugins

* fix: move has_response check to just after config loading

* tests: uncomment all tests, make them more specific

plugins with response method make the whole request buffered, so they
must be segregated if you want to check non-buffered behaviour.

* store response check in ctx, not the plugin iterator

* test for non-buffered request after handling a response phase

* chore(runloop) just use one context value for buffered proxying

Originally we implemented buffered proxying for `2.0` that had PDK function
`kong.service.request.enable_buffering()`. This enabled Lua plugins to read
the response body on `header_filter` phase but did not allow Go plugins to
do so as we cannot yield on `header_filter` and thus that phase is not available
for Go plugins.

We then decided to implement a fake `response` phase. This was implemented but
it added `ngx.ctx.has_response` context value that is not used anywhere else
than detecting whether to do buffered proxying or not. The mentioned `enable_buffering`
function already set `kong.ctx.core.buffered_proxying`. This commit changes so that
we have only one context value to detect that, so both of the above were removed and
now we detect buffered proxying by `ngx.ctx.buffered_proxying`. The reason why not
using `kong.ctx.core` for this is that in hot paths we already have `ngx.ctx` available
and thus it makes sense to use it.

This commit also clear the `ngx.ctx.buffered_proxying` in case it didn't really happen.

Co-authored-by: Aapo Talvensaari <aapo.talvensaari@gmail.com>