fix(exit-transformer) use route to detect "unknown"
The definition of unknown is a kong exit response that is not tied to
anything, like a 404 on a missing service/route, or an internal error.

We used to rely on kong not setting the workspace_id on this particular
context, but it seems that's no longer the case, and it's always the
default.

Can we assume that a request context with no route does not belong to
anything?