fix(acl) deny should not deny proxying if denied group does not match (#6354)
### Summary

It was reported with #6281 that ACL plugin with `config.deny=do-not-deny-anyone`
does not work in all scenarios. I found possible scenarios:

1. There was credential (user did authenticate) but did not have consumer (ldap, openid) nor did not set authenticated groups
2. The user was not authenticated at all (no auth plugins added)
3. The authentication plugin was set to `config.anonymous` and the user didn't authenticate, thus request is `anonymous`

All of this lead to `401`s on `master`. This differs from `2.0` (and `1.4` behavior as reported
by user — mind you there are some differences between versions on `403` vs `401` status codes).

This PR fixes things to behave like `2.0` did with the only difference is
that `2.0` returns `403` in some cases whereas `2.1` returns `401` (the
`403->401` are fixes only in `2.1` series).

### Issues Resolved

Fix #6281