fix(migrations) adjust consistency level for cassandra to avoid duplicate workspaces
### Summary

It has been reported on multiple issues that migrations with Cassandra may lead to
multiple `default` workspaces to be created in database. Most recently by @jeremyjpj0916
on #6403.

This commit tries to tackle the issue by setting consistency of `INSERT` (that adds
default workspace) query to `QUORUM` and consistency of `SELECT` (that tries to determine
if there is a default workspace already) to `SERIAL`.

### Issues Resolved

Fix one part of #6403.