fix(conf) default nginx_user to kong, not nobody
If we install and run Kong with everything default, whenever Kong tries to write something to the Kong prefix, things will break and we will get a permission denied error. This applies to:

- DP nodes trying to write the config.json.gz file to the Kong prefix

- go pluginserver trying to write the go.lua file write to the Kong prefix

- Lua plugins trying to write files to the Kong prefix

And other cases. This is because, by default, the Kong prefix is owned by `root`. When we do a `apt-get install` or `yum install`, the `/usr/local/kong/` prefix (default) is owned by `root` by default. After that, if we do a `kong start` without changing our operating system's `$USER`, the nginx master process will be run as `root` and the nginx workers processes will be run as `nobody`. It means that if we do everything default, the nginx workers processes can’t write anything to the Kong prefix and we’ll get the permission denied errors.

There are a few immediate workarounds for this:

- add the `nginx_user` to the `root` group; or

- change the file system permissions to allow the `nginx_user` the ability to write the cache file to `/usr/local/kong`; or

- allow every user to write to `/usr/local/kong`

All of these workarounds are dangerous and hacky and we don’t document any of these as recommended fixes. We should fix the main cause of the issue. Here are the possibilities:

1) write files somewhere where any user has permissions, like `/tmp`. Maybe even create a `kong_prefix/tmp` folder where it's explicitly said that anyone has permissions over it;

2) set `nginx_user` to `kong` by default (instead of `nobody`), but drop to `nobody` if the `kong` user/group credentials are unavailble. This requires the non-root capability that was implemented at the packaging level [1]. The `nginx_user` configuration will only be changed for environments that have the `kong` user available, so we don't break the source based installs and don't change the default for developers;

3) pre-open all descriptors we need from the master process like nginx seem to do with its logs (e.g. error.log);

4) ship a `kong.conf` with `nginx_user` set to `kong` at the packaging level;

Option 1) is a hack, we’d need to repeat it everywhere we write files in the Kong prefix (Kong codebase, Lua plugins, Go pluginserver, etc). It is also dangerous. The worst option.

Option 2) is a general solution and the one we are going with. Since nginx worker processes will be owned by the `kong` user, and since the default prefix `/usr/local/kong` is now also owned by the `kong` user [1], things that try to write to the Kong prefix won't fail. User can still specify `nobody nobody` if they don't use Hybrid mode or Go plugins and the kong exists on their system and we do not introduce breakages for these users. This may be an obscure use case, but running Kong as nobody is not an uncommon use case. Thanks @dndx for suggesting improvements to this option.

Option 3) is more difficult. Currently there are technical issues that prevent it. We can pre-open all descriptors from the master process in the `init` phase, but it requires passing file descriptors around the codebase. In the `init` phase we have restricted access to some of the important data-sharing mechanisms between workers running in different phases. Also remember that it's not just hybrid mode that is broken, it's also Go plugins and custom Lua plugins that write things to the prefix. Go plugins do not have `init` phases.

Option 4) would mean essentially applying the patch [2] after package installation. Shipping a `/etc/kong/kong.conf.default` with `nginx_user` set to `kong` by default there and also overriding the environment variables. The problem is that the breaking change risk increases if people have `nginx_user` set to something other than the default `nobody`. This is difficult to get it right and it's a breaking change.

The idea of this PR is to fix the permission denied errors when running Kong with the default settings without causing breaking changes by implementing option 2). If the `kong` user does not exist, the user will be dropped to `nobody` as it was before. If the `nginx_user` configuration is set to something else other than `kong` (now the default), we won't change anything and the configuration value will be preserved. If it's set to `nobody` the behavior will be the same as before.

[1] https://github.com/Kong/kong-build-tools/pull/339
[2] https://github.com/Kong/kong/pull/6422/files

Co-authored-by: Datong Sun <datong.sun@konghq.com>