perf(utils) use luajit's table.isarray for array detection
### Summary

OpenResty's fork of LuaJIT has `table.isarray` library that we can use for
checking whether a table is array or not.

This commit also changes the special case of `Lapis` array to another function
called `utils.is_lapis_array` because its implementation was rather opinionated
and it might cause bugs elsewhere where this is not expected.