fix(license) license is relevant in runner
Implemented with a hack in `featureset`. When called through runner,
the global `kong` variable is not properly set yet and setting
`kong.license` has to be done from the license_helpers module
itself (not sure why).

It's not desired to be polling the real data every time in the normal
flow, so we test explicitly for `kong`. If it exists, we know that the
`kong.license` value is the correct one (might be nil for frEE
instances).