fix(migrations) prevent cassandra migrations on creating multiple default workspaces
### Summary

It has been reported on multiple issues that migrations with Cassandra may lead to
multiple default workspaces to be created in database. Most recently by
@jeremyjpj0916 on #6403.

This commit tries to tackle the issue by setting consistency of INSERT (that adds
default workspace) query to QUORUM and consistency of SELECT (that tries to determine
if there is a default workspace already) to SERIAL.

The 212 to 213 uses same migration as 200 to 210 (which we fixed), so we can just
reuse those on 212 to 213 too.

We had in multiple places where we tried to ensure cassandra default workspace.
This commit also moves that to single place.