fix(db) do not export ratelimiting_metrics
### Summary

Rate-limiting metrics is problematic on db exports / imports as
reported by @jeremyjpj0916 on #6310.

There are several issues on that, but it seems like the main issue
is that Cassandra schema specifies the value of ratelimiting_metrics
as `COUNTER`. Also it probably does not make sense to export these
anyway as agreed on that issue.

### Issues Resolved

Fix #6310