fix(go) log-preload data was using module-level scope
Go plugins run the `Log` phase under a timer, which can happen after
the request lifetime, so result of the `kong.log.serialize()` is
preloaded into a module-level table.  Two different Go plugins or two
concurrent requests on the same worker could interfere with each other.

This commit stores this request-specific information using the coroutine
of the timer itself as a key.