perf(migrations) store workspace id after first retrieval
### Summary

On issue #6403 we have tried to chase the reason for why selecting
default workspace id sometimes fail. This commit tries to make it
so that if we ever get workspace id, we will store it in module
level variable for further migrations that might need it, and not
trying to do select each time, which seems flaky according to #6403.

We labeled this as `perf` as that is one thing that it surely does,
but we are not sure if it fixes the #6403, and even if it does, we
still do not fully understand the root cause of it. But it is likely this
change will reduce the frequency at which #6403 occurs.

Co-authored-by: Datong Sun <datong.sun@konghq.com>