fix(openid-connect) add config.cache_tokens_salt
### Summary

Fixes issue of sharing cache key on same credentials but
with different endpoint arguments. Value for this parameter
is auto-generated by default which mean that token endpoint
cache is not automatically shared between different plugin
instances.

This can be considered a breaking change for caching,
but we want the default to be more strict.