feat(redis) connect to redis cluster over ssl/tls
New options 'ssl', 'ssl_verify' and 'server_name' added to the
'connect()' method for connecting to Redis Cluster over TLS.

It reuses the options implemented here:
https://github.com/Kong/kong-ee/pull/2005

But it puts the table in the top level of the `connection()` function so
that it can be passed both to vanilla redis and redis cluster.

Also replace our own redis-cluster library by the upstream one. Our
library was behind the upstream, lacking some features (including this
one) and a few other bug fixes. The two libraries are now in sync.