fix(cli) more atomic reload with dbless
### Summary

On #6464 it was reported that on reload we might affect old workers that
are still processing requests, that were started before reload happened.

This PR tries to make it more atomic by always clearing the shadow caches
only, and loading the config to shadow cache and then flipping the cache.
Previously it was always clearing the current page and loading new config
there, which as the #6464 describes, makes old workers in empty state.

### Issues Resolved

Fix #6464