feat(plugins) minimal and experimental inter subsystem RPC
OpenResty `http` and `stream` subsystem keeps Lua code in separate Lua VM spaces; and for the time being, the shdict mechanism does not yet support inter subsystem access.

This feature adds a simple RPC on top of a datagram-mode unix domain socket. Each stream plugin can optionally register a RPC handler that takes the simplest form - accepts a string and returns a string. Later on the RPC can be made from the `http` subsystem to gather data from the stream subsystem on demand.

This feature is very simple and experimental, we may decide to remove it later once cross subsystem shdict became available.

Co-authored-by: Datong Sun <datong.sun@konghq.com>