feat(jwt-signer) add config.enable_hs_signatures
### Summary

The latest Kong OpenID Connect library disabled HS signature verification
by default for security purposes. This means that any tokens that are signed
by HS256, HS384 or HS512 algorithms cannot be verified anymore by this plugin.
This commit adds a new setting that can be used to turn on HS signatures.
The default setting for this parameter is `false` which means that tokens
signed with shared secret algorithms are not allowed. It is rarely needed
to enable this with this plugin as the HMAC keys, containing the shared
secrets, will then also need to be published in JWKS uri.