feat(openid-connect) add config.enable_hs_signatures (#69)
### Summary

The latest Kong OpenID Connect library disabled HS signature verication
by default for security purposes. This means that any tokens that are
signed by HS256, HS384 or HS512 algorithms cannot be verified anymore
by this plugin (e.g. with `config.client_secret`). This commit adds a
new setting that can be used to turn on HS signatures
(`config.enable_hs_signatures`). The default setting for this parameter
is `false` which means that tokens signed with shared secret algorithms
are not allowed.