chore(*) prevent lua_pack from leaking to globals
### Summary

`lua_pack` is especially bad as it monkey patches the `string` in
Lua. This means that whenever you require the `lua_pack` the `string`
is modified. We don't want this to happen, thus we have this special
construct there. We should replace `lua_pack` or modify `lua_pack`
at some point.