hotfix(schemas) disallow problematic ascii characters in utf8 names
Certain characters (such as /) can be problematic in identifiers used in urls. We have decided to accept all the characters in the utf-8 range *except* that for the ASCII range we allow only certain "url-safe" ones.

The tests that we previously had with Routes and Services have been reinstated to check for these values.