tests(db) add utf-8 tests for consumers
Currently consumers have no restrictions on their username. We can't add typedefs.utf8_names to them because that might break Kong for someone using "invalid" usernames under that typedef (for example using "/" in the username).

So we add tests that ensure that customers *can* be named using utf8, but we *don't* add the tests which would make it fail. This way if in the future we want to restrict consumers away from utf-8 this test will fail.