fix(keyring) add encryption for subschemas (FT-1558)
Previously, with the new implementation of keyring encryption that makes use of DAO transformations, we could not encrypt plugin fields. Only top-level DAO fields from core/plugin entites could. That's because we were not encrypting plugin subschemas. So far we have been only calling `Entity.new` to encrypt top-level core schemas such as DAO core and plugin entities. We were not calling `Entity.new_subschema` to encrypt subschemas.

We are now able to find all encrypted fields from subschemas that are in the plugin fields with the implemented `find_encrypted_fields` algorithm. It's a recursive algorithm that finds all plugin fields that have `encrypted = true`. If at any point during the recursion we find a field of type `record` (nested fields), we recurse over this field to find the `encrypted` fields of this record and add the d "." (dot) character to prefix the nested config field names. For example, we want the transformation input to be `config.client_id`, `config.aws_key` and `config.redis.password`.

A problem that had to be solved was returning the correct data format to the DAO's `Schema:transform()` function upon encrypting/decrypting fields. That is, returning

```
{
  ["config"] = {
    ["client_id"] = "$ke$1$-sPIE42nN-758b2a8f8d5d7f069a5346be-26" } }
```

instead of

```
{ ["config.client_id"] = "$ke$1$-sPIE42nN-758b2a8f8d5d7f069a5346be-26" }
```

That was solved by making use of the `set_field` function that is available in the `/kong/kong/db/schema/init.lua` module. I did not want to expose that function to other modules because that would other changes in that module itself which is constantly being changed by core. In my opinion, it's better to copy and paste it to the `/kong/kong/db/schema/entity.lua` module to avoid merge conflicts. That function should be stable enough that we shouldn't worry about it being changed.

In this commit, subschemas such as plugin config fields that have the field `encrypted = true` now are getting encrypted. This commit is marked as a `fix` because previous versions did not have the issue.